コリジョンとエネミーの雛形作成。リストマネージャ試作

+++ 2024_Project0/supporter.cpp
@@ -6,10 +6,10 @@ D3DXVECTOR3 VectorToAngles(const D3DXVECTOR3& vector)
 	D3DXVECTOR3 angles;
 
 	// Yaw（ヨー）を計算
-	angles.y = atan2(vector.x, vector.z);
+	angles.y = atan2f(vector.x, vector.z);
 
 	// Pitch（ピッチ）を計算
-	angles.x = atan2(vector.y, sqrt(vector.x * vector.x + vector.z * vector.z));
+	angles.x = atan2f(vector.y, sqrt(vector.x * vector.x + vector.z * vector.z));
 
 	// Roll（ロール）は0度に設定
 	angles.z = 0.0f;
@@ -29,9 +29,9 @@ D3DXVECTOR3 AnglesToVector(const D3DXVECTOR3& angles)
 	float pitch = (angles.x);
 
 	// ベクトルを計算
-	vector.x = sin(yaw) * cos(pitch);
-	vector.y = sin(pitch);
-	vector.z = cos(yaw) * cos(pitch);
+	vector.x = sinf(yaw) * cosf(pitch);
+	vector.y = sinf(pitch);
+	vector.z = cosf(yaw) * cosf(pitch);
 
 	return vector;
 }
